Improve value extraction performance
Describing `any` as a String without precaution can cause performance issues when the value's description is very long, like it happens for large arrays or dictionaries for example.
In this case, using a more compact marker is preferable. A common candidate is the hashValue of Hashable values. 
Thus, when `any` is Hashable, the library now uses its `hashValue` instead of the value itself as a path component. In other cases, it uses `any` description.