disable main thread checker for testing
I encounter two problems when trying to test with Xcode 11.3 on macOS 11.4:

- Xcode 11.3's main thread checker crashes when running on macOS 11.4.

- `swift test` fails with the following errors:

  ```
  (dlopen_preflight(/Users/mayoff/TestProjects/swift-case-paths/.build/x86_64-apple-macosx/debug/swift-case-pathsPackageTests.xctest/Contents/MacOS/swift-case-pathsPackageTests): Library not loaded: /usr/lib/swift/libswiftXCTest.dylib
    Referenced from: /Users/mayoff/TestProjects/swift-case-paths/.build/x86_64-apple-macosx/debug/swift-case-pathsPackageTests.xctest/Contents/MacOS/swift-case-pathsPackageTests
    Reason: image not found)
  ```

To work around the problems, this commit turns off the main thread checker for the Test action in the CasePaths scheme.

If on macOS 11.4 and testing with Xcode 11.3, use the IDE, or use `xcodebuild -scheme CasePaths test`, to run the tests.