use destructiveProjectEnumData value witness function (#36)
* use destructiveProjectEnumData value witness function

The destructiveProjectEnumData function in the enum value witness table
transforms an enum value into its associated values in place, by zeroing
out any enum tag bits stored in the payload area of the value.

We can use this function after verifying that the tag of the enum
value is the expected tag. This means we still have to use the
Mirror-and-embed strategy until we can cache the expected tag.

Performance without this commit, on my iMac Pro:

name               time        std        iterations
----------------------------------------------------
Success.Manual       34.000 ns ± 232.27 %    1000000
Success.Reflection 3095.000 ns ±  22.64 %     470183
Failure.Manual       36.000 ns ± 167.71 %    1000000
Failure.Reflection   83.000 ns ± 152.31 %    1000000

Performance with this commit:

name               time       std        iterations
---------------------------------------------------
Success.Manual      35.000 ns ± 158.00 %    1000000
Success.Reflection 167.000 ns ±  70.60 %    1000000
Failure.Manual      37.000 ns ± 197.47 %    1000000
Failure.Reflection  82.000 ns ± 135.63 %    1000000

* Update Sources/CasePaths/EnumReflection.swift

* Add test cases

* make testOptionalPayload pass

I have to explicitly wrap the return value of `associatedValues(of:)` in
a `.some` (even though that return value isn't `Optional`!).

I broke out the calls to `path.extract` in the test case because
`XCTAssertEqual` takes `@autoclosure` arguments, which complicates
stepping in with lldb.

* Remove stale assertion

* handle indirect enums correctly

An indirect enum case stores a pointer to a HeapObject whose payload is
the associated value. Thus the extract function needs to handle an
indirect enum case differently than a direct enum case. This commit
implements proper handling of indirect enum cases.

* macdinking

typeDescriptor should be in EnumMetadata since it returns an EnumTypeDescriptor.

isInhabited and supporting properties are only used when for Swift < 5.2, so conditionalize them accordingly.

* Update Sources/CasePaths/EnumReflection.swift

Co-authored-by: Stephen Celis <stephen.celis@gmail.com>

* Update Sources/CasePaths/EnumReflection.swift

Co-authored-by: Stephen Celis <stephen.celis@gmail.com>

* Update Sources/CasePaths/EnumReflection.swift

Co-authored-by: Stephen Celis <stephen.celis@gmail.com>

* Bring EnumReflection.swift to 100% coverage

The new testIndirectCompoundPayload has two enum cases (one
direct, one indirect) with payloads too large to fit in an existential
container. The test case exercises the isNonInline code paths in
allocateBoxForExistential and deallocateBoxForExistential.

The new testNonEnumExtract verifies that the extract method returns an extractor that always returns nil for a non-enum Root type.

I changed extract to use the ?: operator to choose how to extract the associated value. Without this change, Xcode incorrectly reports some close braces as uncovered.

I've enabled code coverage for the CasePaths scheme and verified that the tests cover every line of EnumReflection.swift.

* add forgotten test

Oops. This was supposed to be included in the prior commit to bring
EnumReflection.swift to 100% coverage.

* Don't use __nonnull

The Linux compiler doesn't like it.

The explicit force-unwrap in EnumReflection.swift isn't strictly necessary because the function's return type is imported as implicitly-unwrapped.

* work around Xcode 11.7 compiler crash

* undo bad change

I forgot to put the test type back after I stopped experimenting with workarounds.

* Update Package.swift

* Update Package@swift-5.1.swift

* remove use of NSObject

On Linux, XCTestCase isn't a subclass of NSObject.

* don't use XCTUnwrap

Xcode 11.3 supports XCTUnwrap in an IDE build, but not in a CLI `swift
test` build (puzzled-face-emoji).

* work around SR-12044 properly

I broke the SR-12044 workaround. I misunderstood the old code and so reimplemented it incorrectly when I changed how we access metadata.

Co-authored-by: Rob Mayoff <git@rob.dqd.com>
Co-authored-by: Stephen Celis <stephen.celis@gmail.com>
Co-authored-by: Stephen Celis <stephen@stephencelis.com>