fix issue #40 (case initializer promoted to return Optional) (#41)
* assert if issue #40 is detected in a DEBUG build

Swift promotes (A) -> B to (A) -> B? if necessary to make a statement
type-check. This can lead to an enum case initializer being promoted and
making extract erroneously return nil. Fixing it is hard, so instead I'm
detecting the problem and asserting.

* remove unneeded else if clause

Leftover from trying to do the right thing instead of just asserting.

* fix grammar

* handle optionals properly

fixes #40 even better

* add test case using embedded tag bits

The enum E is stored in a single word. The high bit of
the word is 0 for the c1 case and 1 for the c2 case.

Co-authored-by: Rob Mayoff <git@rob.dqd.com>