Breaking change to CasePath.modify
We'd like to make a breaking change to `CasePath.modify`:

```diff
-func modify<R>(inout Enum, (inout Case) throws -> R) throws -> R
+func modify(inout Enum, (inout Case) throws -> Void) rethrows -> R
```

1. It no longer throws (unless the transform closure throws)
2. It no longer returns a result through the continuation.

For 1., we're not sure error handling in the signature makes sense.
Modifying the wrong case seems more like a programmer error than
anything else. So instead, we'll runtime warn.

For 2., losing `throws` means the result would need to be optionalized.
While this is possible, we also don't think values are typically
assigned through the result of this closure, so let's remove support for
this completely.

For most folks we believe this change will only introduce a build
warning:

> **Warning:** No calls to throwing functions occur within 'try' expression

So hopefully not too disruptive!