Introduce `CasePathProtocol`
`CasePathProtocol` is an alternative to an `AnyCasePath` holding
escaping closures for `embed` and `extract`. It allows case path logic
to be better optimized by the compiler, including inlining.

Internal algorithms for appending case paths has been rewritten to take
advantage of case path conformances, through `@CasePathable` will
continue to generate closure-based `AnyCasePath`s for now.

We can consider adding a configurable option in a minor release if we
want things to be truly backwards compatible, even with macro-generated
code, or we can defer generating these more performant case paths in a
2.0.