Add CasePathsCore usage example
I was looking for an explanation on how to use  `CasePathsCore` and happened to found it in the `CasePathable` documentation comment. 

Before I found it though I wasn't aware that `CasePathable` was the protocol the macro synthesized conformance for. This moves the information to a place that is easier to find. Although maybe it should even be in a CasePathsCore README?